add auth component
