+++ src/util/pd-api-wrapper.js
@@ -1,6 +1,6 @@
 /* eslint-disable no-await-in-loop */
 /* eslint-disable no-loop-func */
-import PDOAuth from 'util/pdoauth';
+
 import {
   api,
 } from '@pagerduty/pdjs';
@@ -8,7 +8,7 @@ import axios from 'axios';
 import Bottleneck from 'bottleneck';
 
 import {
-  PD_OAUTH_CLIENT_ID, PD_OAUTH_CLIENT_SECRET, PD_USER_TOKEN,
+  PD_USER_TOKEN,
 } from 'config/constants';
 import {
   compareCreatedAt,
@@ -30,9 +30,9 @@ export const getPdAccessTokenObject = () => {
     tokenType = 'bearer';
   }
 
-  // Begin OAuth workflow if env token not present
+  // If the token is not present, something is wrong
   if (!token) {
-    PDOAuth.login(PD_OAUTH_CLIENT_ID, PD_OAUTH_CLIENT_SECRET);
+    return null;
   }
 
   // Return object neeed for PD API helpers
